Add deduplication of watched objects.
This should stop recursion of a watch object that we update the status
of in the reconciler, which would then invoke a new run.

The dedup only allows running of a reconcile against a uniq object at a
time but will adhere to the concurrency if there are several different
objects at the same time.